requirements.txt
asgiref==3.4.1
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
decorator==5.1.0
Django==3.2.13
django-admin-interface==0.17.3
django-colorfield==0.4.5
django-cors-headers==3.10.0
django-daterangefilter==1.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-simple-history==3.0.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
ipython==7.30.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
numpy==1.21.4
openpyxl==3.0.9
pandas==1.4.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.6.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pycountry==20.7.3
Pygments==2.10.0
pytest==7.1.2
python-dateutil==2.8.2
pytrials==0.2.0
pytz==2021.3
regex==2021.11.10
requests==2.26.0
Shapely==1.8.0
six==1.16.0
soupsieve==2.3.1
sqlalchemy==1.4.36
sqlparse==0.4.2
tenacity==8.0.1
tqdm==4.62.3
traitlets==5.1.1
urllib3==1.26.7
utils==1.0.1
wcwidth==0.2.5